dev-t2: union thrillio project and restful project

+++ src/main/java/org/websitebook/ws/dao/ReadableDAO.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface ReadableDAO<T> {
 	
-	T getById(int id) throws SQLException;
+	T getById(Long id) throws SQLException;
 	
 	List<T> getAll() throws SQLException;
